Fix typescript issue
